More reliable future parallellization
When using the future parallellization framework, future is able to
automagically detect variables that are not explicitly passed to the
environment (globals). This is great for interactive programming,
as "things usually just work".

When using the Seurat 3.0 package with parallellized FindAllMarkers,
I found some warnings stating that the larger clusters were skipped
because there were some global variables that were too big (the
seurat object), and I was suggested to change
`options(future.globals.maxSize=)` to a large
enough value.

This commit explicitely passes all the "global" used objects
to the parallellized functions to ensure we don't hit the global
variable size limit.